+++ src/main/java/com/example/demo/web/controller/PermissionController.java
@@ -71,7 +71,7 @@ public class PermissionController {
 
     @GetMapping(value = "/delete")
     @ResponseBody
-    public Object delete(Integer[] id) {
+    public Object delete(Long[] id) {
         permissionService.delete(id);
         return ResponseEntity.buildOk();
     }
